program: Harden transient stake account checks during update
#### Problem

Because of the permissionless pool updating functions of the stake pool
program, it's possible for malicious users to hijack stake accounts
owned by the pool. If someone hijacks a transient stake account and
reassigns it to the pool's withdraw authority, it's possible for the
pool to wrongly account for the lamports in the hijacked stake account.

#### Summary of changes

Harden the accounting logic for transient stakes during update. This
handles a few edge cases.

First, ignore any transient stake account delegated to the incorrect
validator. This is a simple check to filter out potential failures
later.

Next, take proper action based on the `StakeStatus` for the validator:
* if `Active`, do the same thing as before
* if `DeactivatingTransient` or `DeactivatingAll`, mostly do the same
  thing as before, but ignore if it's not deactivating or inactive
* if `ReadyForRemoval` or `DeactivatingValidator`, ignore