program: Protect against cluster-restart hijacks
#### Problem

During cluster restart in which a stake account is reset to
`Initialized`, the stake pool will properly merge those funds back into
the pool, but does not update the validator status.

This means it's possible for exploiters to take control of a stake
account and have it be counted by the stake pool.

#### Summary of changes

This includes two fixes:

* only count active validator stakes if they're usable by the pool,
  similar to how transient stakes are handled
* update the validator status correctly after merging the inactive stake
  into the reserve