program: Don't remove validator if any lamports present
#### Problem

In some situations where the transient account is hijacked and a
validator has been marked for removal, the stake pool incorrectly counts
the transient stake as being part of the pool.

During the cleanup phase, the transient lamports immediately get wiped
out, causing a share price to drop. Malicious actors can profit from
knowing when these spikes will happen.

#### Summary of changes

This is not the full fix to the issue, but as a start, to avoid spikes
in the pool token value by incorrectly wiping out a validator with
lamports, simply check to make sure that there are no more lamports
delegated to the validator before cleaning up the entry.

If a stake pool does enter this state, another update will be required
to deactivate the transient stake so that the lamports can be reclaimed.