program: Prevent removing validator with transient lamports
#### Problem

During `WithdrawStake`, it's possible to remove a validator stake
account if all the stake accounts are at minimum. However, if that
particular validator has a transient stake account, then those lamports
are inadvertently removed from the pool at the same time, essentially
lost and no longer usable by the pool.

#### Summary of changes

Add an additional check to not remove a validator if it has any
transient lamports associated with it.