program: Reset preferred validator during withdrawal
#### Problem

During `WithdrawStake`, it's possible to completely remove stake
accounts if all validator stake accounts are at a minimum amount. If a
preferred validator is removed in this way, the preferred validator is
not reset, making it impossible for any other withdrawal to happen.

#### Summary of changes

Reset the preferred validator during withdraw-stake if needed. At the
same time, don't brick all withdrawals if the preferred validator is not
found for some reason, and proceed either way.