program: Set status to `DeactivatingAll` if transient lamports
#### Problem

During a cluster restart situation, or a situation in which stake
accounts can be modified outside of the stake protocol, there can be a
situation in which a transient stake account is in the `Initialized`
state, but there are transient lamports accounted for by the pool.

In the current logic, the status would be wrongly set to
`DeactivatingValidator`, even though there exists a stake account with
transient lamports associated.

#### Summary of changes

Always set the status to `DeactivatingAll` if transient lamports exist
on a given validator.