program: Restrict stake check during withdrawal to active
#### Problem

During withdrawal, we check that there's no possible validator with
enough active stake to service the withdrawal, but we also prevent
withdrawing from a stake account that's marked for removal.

This means it's possible to have a situation where the only withdrawal
candidate is being removed, making it impossible to withdraw.

#### Summary of changes

Restrict the stake check during withdraw-stake to only include active
valdiators.